Pager: 抽取公共方法
